+++ CriteoPublisherSdk/Sources/CR_CdbBid.m
@@ -74,8 +74,6 @@ static CR_CdbBid *emptyBid;
     _creative = creative;
     _ttl = ttl;
     _displayUrl = displayUrl;
-    _dfpCompatibleDisplayUrl = [NSString cr_dfpCompatibleString:displayUrl];
-    _mopubCompatibleDisplayUrl = [NSString cr_mopubCompatibleDisplayUrlForDisplayUrl:displayUrl];
     _insertTime = insertTime;
     _nativeAssets = [nativeAssets copy];
     _impressionId = impressionId;